sample sort prince

+++ hw6/hw6helper.h
@@ -4,10 +4,10 @@
 
 /*Quick sort implementation from the stack ooverflow:
 https://stackoverflow.com/questions/19894206/quicksort-with-double-values*/
-void quicksort(double a[], int n) {
+void quicksort(int a[], int n) {
     if (n <= 1) return;
-    double p = a[n/2];
-    double b[n], c[n];
+    int p = a[n/2];
+    int b[n], c[n];
     int i, j = 0, k = 0;
     for (i=0; i < n; i++) {
         if (i == n/2) continue;
@@ -38,7 +38,7 @@ long binSearch(double *v, long vlen ,double item){
 	return -1;
 }
 
-void sampleSplits(double *v, double *s, long vlen, long slen){
+void sampleSplits(int *v, int *s, long vlen, long slen){
 	/*Sample every vlen/slen elements*/
 	long interval = (vlen-1)/slen;
 	for(long i = 1; i <= slen; i++){
